Split endpoints into seperate router files
